move NormalizedProtobufMessageName to interchange crate
