+++ src/ore/Cargo.toml
@@ -34,7 +34,6 @@ lazy_static = "1.4.0"
 openssl = { version = "0.10.38", features = ["vendored"], optional = true  }
 pin-project = "1"
 prometheus = { git = "https://github.com/MaterializeInc/rust-prometheus.git", default-features = false, optional = true }
-serde = { version = "1.0.130", features = ["derive"] }
 smallvec = { version = "1.7.0", optional = true }
 stacker = { version = "0.1.14", optional = true }
 structopt = { version = "0.3.25", optional = true }
